backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.55.3
python-dotenv==1.0.0
pydantic==2.5.2
alembic==1.12.1
python-jose[cryptography]==3.3.0
pydantic[email]
pydantic_settings